modularized code base
